試題1_Crawler/requirements.txt
selenium>=4.11.2
psycopg2-binary==2.9.11
sqlalchemy==2.0.47
APScheduler==3.11.2
python-dotenv==1.2.1
ddddocr==1.5.6
requests==2.32.5
pillow==12.1.1
fastapi==0.134.0
uvicorn==0.41.0